[#3051] Process new PR feedback
