+++ src/openforms/scss/components/admin/_datetime.scss
@@ -3,7 +3,7 @@
   gap: 3px;
   padding-right: 2px;
 
-  i {
+  .fa-solid {
     margin-top: 2px;
     cursor: pointer;
   }
